src/main/java/PolishRecord.java
import java.io.BufferedReader;
import java.io.FileReader;
import java.io.FileWriter;
import java.io.IOException;
import java.util.ArrayDeque;

public class PolishRecord {

    private static final char MINUS = '-';
    private static final char PLUS = '+';
    private static final char MULTIPLY = '*';
    private static final char DIVIDED = '/';
    private static final char SPACE = ' ';
    private static final char LEFT_BRACKET = '(';
    private static final char RIGHT_BRACKET = ')';


    public String sourceExpression;
    public String decodedExpression;
    public String name;
    private String[] s;
    public String Answer;
  //  public Double answer;



    public PolishRecord() {

    }


    public PolishRecord(String sourceExpression) {

        this.sourceExpression = sourceExpression;
    }

    /*




     */
    public String decodedExpression(String sourceExpression) {
        StringBuffer polishRecord = new StringBuffer();
        //String polishRecord = "";
        ArrayDeque<Character> stack = new ArrayDeque<Character>();
        int priority;
        for (int i = 0; i < sourceExpression.length(); i++) {
            priority = GetPriority(sourceExpression.charAt(i));

            if (priority == 0) {
                //polishRecord += sourceExpression.charAt(i);
                polishRecord.append(sourceExpression.charAt(i));
            }
            if (priority == 1) {
                stack.push(sourceExpression.charAt(i));
            }
            if (priority > 1) {

               // polishRecord += SPACE;
                polishRecord.append(SPACE);

                while (!stack.isEmpty()) {
                    if (GetPriority(stack.peek()) >= priority) {
                       // polishRecord += stack.pop();
                        polishRecord.append(stack.pop());
                    } else break;
                }
                stack.push(sourceExpression.charAt(i));
            }
            if (priority == -1) {

               // polishRecord += SPACE;
                polishRecord.append(SPACE);
                while (GetPriority(stack.peek()) != 1) {
                    polishRecord.append(stack.pop());
                    //polishRecord += stack.pop();
                }
                stack.pop();
            }
        }
        while (!stack.isEmpty()) {
           // polishRecord += stack.pop();
            polishRecord.append(stack.pop());
        }
        return polishRecord.toString();
    }




    public double calculationResult(String decodedExpression) {
        //String operand = new String();
        StringBuffer br = new StringBuffer();
        ArrayDeque<Double> stack = new ArrayDeque<Double>();
        for (int i = 0; i < decodedExpression.length(); i++) {
            if (decodedExpression.charAt(i) == SPACE) {
                continue;
            }
            if (GetPriority(decodedExpression.charAt(i)) == 0) {
                while (decodedExpression.charAt(i) != SPACE && GetPriority(decodedExpression.charAt(i)) == 0) {
                   // operand += decodedExpression.charAt(i++);
                    br.append(decodedExpression.charAt(i++));

                    if (i == decodedExpression.length()) {
                        break;
                    }
                }

                stack.push(Double.parseDouble(br.toString()));
                br = new StringBuffer();
            }

            if (GetPriority(decodedExpression.charAt(i)) > 1) {
                double a = stack.pop(), b = stack.pop();

                if (decodedExpression.charAt(i) == PLUS) {
                    stack.push(b + a);
                }
                if (decodedExpression.charAt(i) == MINUS) {
                    stack.push(b - a);
                }
                if (decodedExpression.charAt(i) == MULTIPLY) {
                    stack.push(b * a);
                }
                if (decodedExpression.charAt(i) == DIVIDED) {
                    stack.push(b / a);
                }

            }

        }
        return stack.pop();
    }


    private static int GetPriority(char a) {
        switch (a) {
            case MULTIPLY:
            case DIVIDED:
                return 3;
            case PLUS:
            case MINUS:
                return 2;
            case LEFT_BRACKET:
                return 1;
            case RIGHT_BRACKET:
                return -1;
            default:
                return 0;
        }
    }


    public String ReadToFile(String name) {

        try {
            BufferedReader reader = new BufferedReader(new FileReader(name));
            String str;
            while ((str = reader.readLine()) != null) {
                if (!str.isEmpty()) {
                    sourceExpression=str;
                }
            }
        } catch (Exception e) {
            e.getMessage();
        }
        return sourceExpression;
    }




    public String[] WriteToFile(String name, String [] s) {
        try {

            FileWriter writer = new FileWriter(name);
            for(int i=0;i<s.length; i++) {
                writer.write(s[i]);
                writer.append('\n');
                writer.flush();
            }

        } catch (IOException e) {
            e.printStackTrace();
        }
        return s;
    }




    public  String CorrectExpression(String s ){
         StringBuilder sb = new StringBuilder(s.length());
         char a='+';
         char b='-';
         char c='*';
         char d='/';
         char e='(';
         char f=')';
        for(int i = 0; i < s.length(); i++){
            char p = s.charAt(i);
            if(p==a || p==b || p==c || p==d || p==e || p==f){
                continue;
            }

            if(p <47 || p >58){

                sb.append(p);
           }
        }
        return sb.toString();
    }

}